refactor aesthetics
