+++ src/typing.ts
@@ -1,10 +1,14 @@
-import type {
-  RootChannel, OpChannel, ColorChannel,
-  ConstantChannel, LambdaChannel, Labelcall,
-  URLLabels, Labelset, 
-} from './shared.d'
+
 
 import type * as DS from './shared.d'
+const transforms = new Set(['linear', 'sqrt', 'log', 'literal'])
+
+export function isTransform(input: unknown): input is DS.Transform {
+  if (typeof input === 'string' && transforms.has(input)) {
+    return true
+  }
+  return false;
+}
 
 export function isOpChannel(input: DS.ChannelType): input is DS.OpChannel<any> {
   return input['op'] !== undefined;
@@ -20,10 +24,10 @@ export function isConstantChannel(
   return input['constant'] !== undefined
 }
 
-export function isURLLabels(labels: Labelcall): labels is URLLabels {
-  return labels !== null && (labels as URLLabels).url !== undefined;
+export function isURLLabels(labels: DS.Labelcall): labels is DS.URLLabels {
+  return labels !== null && (labels as DS.URLLabels).url !== undefined;
 }
 
-export function isLabelset(labels: Labelcall): labels is Labelset {
-  return labels !== null && (labels as Labelset).labels !== undefined;
+export function isLabelset(labels: DS.Labelcall): labels is DS.Labelset {
+  return labels !== null && (labels as DS.Labelset).labels !== undefined;
 }
